Moved creation of Jinja environment into dynamic provider base class, and changed JinjaEnvironment class into a file of functions
