+++ pulumi_snowflake/database/database_provider.py
@@ -1,7 +1,6 @@
 from pulumi_snowflake import Client
 
 from ..baseprovider import BaseDynamicProvider
-from ..baseprovider.jinja_environment import JinjaEnvironment
 from ..provider import Provider
 
 
@@ -13,8 +12,8 @@ class DatabaseProvider(BaseDynamicProvider):
     def __init__(self, provider_params: Provider, connection_provider: Client):
         super().__init__(provider_params, connection_provider, "DATABASE", [])
 
-    def _generate_sql_create_statement(self, attributesWithValues, validated_name, inputs):
-        template = JinjaEnvironment.create_template(
+    def _generate_sql_create_statement(self, attributesWithValues, validated_name, inputs, environment):
+        template = environment.from_string(
 """CREATE{% if transient %} TRANSIENT{% endif %} DATABASE {{ full_name }}
 {% if share %}FROM SHARE {{ share | sql_identifier }}{% endif %}
 {% if data_retention_time_in_days %}DATA_RETENTION_TIME_IN_DAYS = {{ data_retention_time_in_days | sql }}{% endif %}
